Break out qualified name refs
